Fix: Input data Preprocessing
fp16 연산 여부가 configuration에 명시되어 있으므로 해당 value에 따라 16bits 혹은 32bits 연산을 선택적으로 수행해야하는데 중간에 일괄적으로 16bits로 이미지 텐서를 변경하는 부분 처리

+++ detectors/hrnet.py
@@ -62,7 +62,7 @@ class HRNet(nn.Module):
                 (int(self.img_size[0]), int(self.img_size[1])),
                 flags=cv2.INTER_LINEAR)
             # hwc -> 1chw
-            model_input = self.pose_transform(model_input).half()  # .unsqueeze(0)
+            model_input = self.pose_transform(model_input)  # .unsqueeze(0)
             model_inputs.append(model_input)
 
         # n * 1chw -> nchw
